Fix ineffective cookie banner dynamic import
