+++ client/blocks/cookie-banner/is-server.ts
@@ -1,0 +1,1 @@
+export const isServer = typeof document === 'undefined';
